S2 frameworks laid out
this is a big commit because now I know exactly how I will implement
multiple currencies.  The Configuration.xml file is how I will
allow people to assign any currency to the game and have any manipulator
assigned to each of the currencies independently.

It's a lot of work in the configuration, but it will pay off with not having
the same limitations that SEc 1 does.